refactor(plan_SERVERSPLIT slice 5): the progressive-cloning bodies leave server.py
The ten `progressive-cloning` tool bodies move VERBATIM out of
`embedded/server.py` into `embedded/tool_sections/progressive_cloning.py`. No
behaviour change, no rename, no tool added or removed: the served surface is
byte-identical to `tests/goldens/tool_surface.json`, and all three loaded
identities of `server.py` still build a full 94-tool app.

This is the first GOLDEN-BACKED section, which is the property the slice exists
to test. Its two goldens — `tests/goldens/progressive_expand_styles.json` and
`progressive_list_stored_elements.json` — belong to `progressive_element_cloner`,
the adapter one layer BELOW the tool bodies, so a pure move cannot reach them.
`tests/test_cloner_schemas.py` is re-run to prove it and `git status` shows both
files untouched: goldens neither moved nor changed in this commit.

`expand_children`'s two `{"error": ...}` returns are carried, not authored, here.
Converting them would be a behaviour change, and plan_SERVERSPLIT §8 forbids one
inside a slice whose entire verification is a byte-identical surface.

server.py: 2839 -> 2660 LOC; the GRANDFATHER cap ratchets DOWN to that measured
post-`ruff format` actual, and `tests/source_scan.py`'s floor ratchets UP 6 -> 7.
The one-line `progressive_element_cloner` alias import server.py was the last
consumer of is pruned with the bodies.

Verified: unit lane 2332 passed / 1 skipped (one fewer than slice 4 for the same
reason: the migration alias-identity pin enumerates `tool_runtime.__all__` and
`dir(server)` and loses a parametrize case with every pruned alias); wire surface
IDENTICAL; three-identity load SECTION_TOOLS 94 / per-app [94, 94, 94];
`--list-sections` subprocess "Total: 94 tools"; ruff format+check, budgets,
suppression owners, vulture and `gen_release_contract --check` clean; ty exit 0
at 78 warnings, one more than the 77 baseline — `expand_children`'s bare
`depth_range: list` is part of the WIRE surface and so cannot be narrowed by a
pure move, and it becomes visible only because `server.py` is in ty's exclude
list while a section module is not. Real Chrome: `test_e2e_data_tools.py`
5 passed; real stdio: `test_e2e_transport.py` 1 passed.
