+++ pyproject.toml
@@ -256,6 +256,13 @@ ban-relative-imports = "all"
 # is the exact subset of server.py's file-wide list the moved lines fire —
 # notably A002 for export_debug_logs' `format` parameter, which is part of the
 # WIRE surface and so cannot be renamed by a pure move. Carried verbatim.)
+"src/stealth_chrome_devtools_mcp/embedded/tool_sections/progressive_cloning.py" = [
+    "ANN401", "E501", "TID251",
+]  # plan_SERVERSPLIT slice 5 (the progressive-cloning bodies moved VERBATIM out
+# of server.py; the exact subset of that file's file-wide list they fire —
+# TID251 the `typing.Any` annotations, ANN401 expand_children's `max_count:
+# Any | None`, which is part of the WIRE surface and so cannot be narrowed by a
+# pure move, E501 four docstring lines. Carried verbatim, nothing widened.)
 "src/stealth_chrome_devtools_mcp/embedded/tool_sections/dynamic_hooks.py" = [
     "E501", "PLR0913", "TID251",
 ]  # plan_SERVERSPLIT slice 4 (the dynamic-hooks bodies moved VERBATIM out of
