+++ src/stealth_chrome_devtools_mcp/embedded/server.py
@@ -53,7 +53,6 @@ from stealth_chrome_devtools_mcp.embedded.tool_runtime import (
     file_based_element_cloner,
     in_memory_storage,
     network_interceptor,
-    progressive_element_cloner,
     response_handler,
 )
 from stealth_chrome_devtools_mcp.embedded.tool_sections import SECTION_MODULES
@@ -1751,184 +1750,6 @@ async def clone_element_complete(
             "url": getattr(tab, "url", "unknown"),
         },
     )
-
-
-@section_tool("progressive-cloning")
-async def clone_element_progressive(
-    instance_id: str, selector: str, include_children: bool = True
-) -> dict[str, Any]:
-    """
-    Clone element progressively - returns lightweight base structure with element_id.
-
-    Args:
-        instance_id (str): Browser instance ID.
-        selector (str): CSS selector for the element.
-        include_children (bool): Whether to extract child elements.
-
-    Returns:
-        Dict[str, Any]: Base structure with element_id for progressive expansion.
-    """
-    tab = await _require_tab(browser_manager, instance_id)
-    return await _with_cdp_timeout(
-        progressive_element_cloner.clone_element_progressive(
-            tab, selector, include_children
-        ),
-        instance_id=instance_id,
-    )
-
-
-@section_tool("progressive-cloning")
-async def expand_styles(
-    element_id: str,
-    categories: list[str] | None = None,
-    properties: list[str] | None = None,
-) -> dict[str, Any]:
-    """
-    Expand styles data for a stored element.
-
-    Args:
-        element_id (str): Element ID from clone_element_progressive().
-        categories (Optional[List[str]]): Style categories to include (layout, typography, colors, spacing, borders, backgrounds, effects, animation).
-        properties (Optional[List[str]]): Specific CSS property names to include.
-
-    Returns:
-        Dict[str, Any]: Filtered styles data.
-    """
-    return progressive_element_cloner.expand_styles(element_id, categories, properties)
-
-
-@section_tool("progressive-cloning")
-async def expand_events(
-    element_id: str, event_types: list[str] | None = None
-) -> dict[str, Any]:
-    """
-    Expand event listeners data for a stored element.
-
-    Args:
-        element_id (str): Element ID from clone_element_progressive().
-        event_types (Optional[List[str]]): Event types or sources to include (click, react, inline, addEventListener).
-
-    Returns:
-        Dict[str, Any]: Filtered event listeners data.
-    """
-    return progressive_element_cloner.expand_events(element_id, event_types)
-
-
-@section_tool("progressive-cloning")
-async def expand_children(
-    element_id: str, depth_range: list | None = None, max_count: Any | None = None
-) -> dict[str, Any]:
-    """
-    Expand children data for a stored element.
-
-    Args:
-        element_id (str): Element ID from clone_element_progressive().
-        depth_range (Optional[List]): [min_depth, max_depth] range to include.
-        max_count (Optional[Any]): Maximum number of children to return.
-
-    Returns:
-        Dict[str, Any]: Filtered children data.
-    """
-    if isinstance(max_count, str):
-        try:
-            max_count = int(max_count) if max_count else None
-        except ValueError:
-            return {"error": f"Invalid max_count value: {max_count}"}
-
-    if isinstance(depth_range, list):
-        try:
-            depth_range = [int(x) if isinstance(x, str) else x for x in depth_range]
-        except ValueError:
-            return {"error": f"Invalid depth_range values: {depth_range}"}
-
-    depth_tuple = tuple(depth_range) if depth_range else None
-
-    result = progressive_element_cloner.expand_children(
-        element_id, depth_tuple, max_count
-    )
-    return response_handler.handle_response(result, f"expand_children_{element_id}")
-
-
-@section_tool("progressive-cloning")
-async def expand_css_rules(
-    element_id: str, source_types: list[str] | None = None
-) -> dict[str, Any]:
-    """
-    Expand CSS rules data for a stored element.
-
-    Args:
-        element_id (str): Element ID from clone_element_progressive().
-        source_types (Optional[List[str]]): CSS rule sources to include (inline, external stylesheet URLs).
-
-    Returns:
-        Dict[str, Any]: Filtered CSS rules data.
-    """
-    return progressive_element_cloner.expand_css_rules(element_id, source_types)
-
-
-@section_tool("progressive-cloning")
-async def expand_pseudo_elements(element_id: str) -> dict[str, Any]:
-    """
-    Expand pseudo-elements data for a stored element.
-
-    Args:
-        element_id (str): Element ID from clone_element_progressive().
-
-    Returns:
-        Dict[str, Any]: Pseudo-elements data (::before, ::after, etc.).
-    """
-    return progressive_element_cloner.expand_pseudo_elements(element_id)
-
-
-@section_tool("progressive-cloning")
-async def expand_animations(element_id: str) -> dict[str, Any]:
-    """
-    Expand the animations slice (schema v2) and fonts for a stored element.
-
-    Args:
-        element_id (str): Element ID from clone_element_progressive().
-
-    Returns:
-        Dict[str, Any]: `animations` — the same schema-v2 object
-        extract_element_animations() documents — plus `fonts`.
-    """
-    return progressive_element_cloner.expand_animations(element_id)
-
-
-@section_tool("progressive-cloning")
-async def list_stored_elements() -> dict[str, Any]:
-    """
-    List all stored elements with their basic info.
-
-    Returns:
-        Dict[str, Any]: List of stored elements with metadata.
-    """
-    return progressive_element_cloner.list_stored_elements()
-
-
-@section_tool("progressive-cloning")
-async def clear_stored_element(element_id: str) -> dict[str, Any]:
-    """
-    Clear a specific stored element.
-
-    Args:
-        element_id (str): Element ID to clear.
-
-    Returns:
-        Dict[str, Any]: Success/error message.
-    """
-    return progressive_element_cloner.clear_stored_element(element_id)
-
-
-@section_tool("progressive-cloning")
-async def clear_all_elements() -> dict[str, Any]:
-    """
-    Clear all stored elements.
-
-    Returns:
-        Dict[str, Any]: Success message.
-    """
-    return progressive_element_cloner.clear_all_elements()
 
 
 @section_tool("file-extraction")
