+++ src/stealth_chrome_devtools_mcp/embedded/tool_sections/__init__.py
@@ -31,6 +31,7 @@ from stealth_chrome_devtools_mcp.embedded.tool_sections import (
     cookies_storage,
     debugging,
     dynamic_hooks,
+    progressive_cloning,
     tabs,
 )
 
@@ -44,4 +45,10 @@ from stealth_chrome_devtools_mcp.embedded.tool_sections import (
 #: mid-migration ``SECTION_TOOLS`` lists the MOVED sections first; nothing reads
 #: that ordering (``release_evidence`` and the contract generator both sort, and
 #: ``--list-sections`` only totals), and it settles once every section has moved.
-SECTION_MODULES = (cookies_storage, debugging, tabs, dynamic_hooks)
+SECTION_MODULES = (
+    cookies_storage,
+    debugging,
+    tabs,
+    progressive_cloning,
+    dynamic_hooks,
+)
